PrinceanWorlds multiple gravity vectors
Princean worlds can now affect and be affected by multiple other objects simoultaneously dependant on ones choice of array size.. My visbile vector was also implemented to further visualize the pull, always showing which planets are interacting.

The vector is left dead in space if the gravity connection is broken. This was not done on purpose. I find it so aesthetic though! A marker showing the death of their connection.