TrueGravity simulation
Finally made the TrueGravity script. Gravity now falls of linearly with distance. I tried to simulate a solar system within a scene. I could not get bodies of moderate size distances to maintain of semblance of a stable orbit. One either has to have bodies of equal size or of widely different.